% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_list.R
\name{bootstrap_list}
\alias{bootstrap_list}
\title{Bootstrap the Data Frames in a List}
\usage{
bootstrap_list(x, n = 1000, rows = 1000, seed = NULL, dfidx = FALSE, ...)
}
\arguments{
\item{x}{The original data frame.}

\item{n}{The number of bootstrapped data frames to return.}

\item{rows}{The number of rows to sample from the original data frame.}

\item{seed}{The seed to use for the random number generator.
Default is NULL.}

\item{dfidx}{If TRUE, use the dfidx::dfidx function to transform into mlogit
data format, i.e., a dataframe with indexes. Default is FALSE.}

\item{...}{Additional arguments to pass to dfidx::dfidx.}
}
\value{
A list of data frames.
}
\description{
This function takes a data frame and returns a list of data frames that are
bootstrapped from the original data frame.
}
\examples{
df <- data.frame(no = seq(length(state.abb)), stabb = state.abb)
bootstrap_list(df)

}
