% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{identity}
\alias{identity}
\title{Identity-ranking data analyzed in Atsusaka and Kim (2025)}
\format{
## `identity`
A data frame with 1,082 rows and 16 columns:
\describe{
  \item{app_identity}{
    Full ranking profile for the main identity ranking question.
  }
  \item{party}{
    Marginal ranking for party (main identity ranking question).
  }
  \item{religion}{
    Marginal ranking for religion (main identity ranking question).
  }
  \item{gender}{
    Marginal ranking for gender (main identity ranking question).
  }
  \item{race}{
    Marginal ranking for race (main identity ranking question).
  }
  \item{anc_identity}{Full ranking profile for the anchor ranking question.}
  \item{household}{Marginal ranking for household (anchor question).}
  \item{neighborhood}{
    Marginal ranking for neighborhood (anchor question).
  }
  \item{city}{Marginal ranking for city (anchor question).}
  \item{state}{Marginal ranking for state (anchor question).}
  \item{anc_correct_identity}{
    Whether the respondent answered the anchor questions correctly.
    This is a binary variable that 1 if the respondent correctly answers
    the anchor ranking question and 0 if otherwise.
  }
  \item{app_identity_recorded}{
    Recorded responses for the main identity ranking question.
  }
  \item{anc_identity_recorded}{
    Recorded responses for the anchor ranking question.
  }
  \item{app_identity_row_rnd}{
    The order in which the items were randomly presented for the
    respondent in the main ranking question.
  }
  \item{anc_identity_row_rnd}{
    The order in which the items were randomly presented for the
    respondent in the anchor ranking question.
  }
  \item{s_weight}{Survey weight.}
}
}
\source{
<https://github.com/sysilviakim/ranking_error>
}
\usage{
identity
}
\description{
Full dataset from survey on relative partisanship used in
Atsusaka, Yuki, & Kim, Seo-young Silvia (2025).
Addressing Measurement Errors in Ranking Questions for the Social Sciences.
Political Analysis, 33(4), 339-360.
https://doi.org/10.1017/pan.2024.33

This data contains Americans' rankings of four sources of their identity,
including political party, religion, gender, and race, for
1,082 respondents.
The columns consist of marginal rankings to the main identity ranking
question and the corresponding anchor question, as well as whether they have
answered the anchor questions "correctly." The anchor ranking question is
used to estimate the proportion of random responses and
to correct for measurement error bias.
}
\keyword{datasets}
