% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{identity_w}
\alias{identity_w}
\title{Identity-ranking data with estimated weights based on inverse probability
weighting}
\format{
## `identity_w`
A data frame with 1,082 rows and 18 columns:
\describe{
  \item{weights}{Estimated weights based on inverse probability weighting.}
  \item{s_weight}{Survey weight.}
  \item{app_identity}{
    Full ranking profile for the main identity ranking question.
  }
  \item{party}{
    Marginal ranking for party (main identity ranking question).
  }
  \item{religion}{
    Marginal ranking for religion (main identity ranking question).
  }
  \item{gender}{
    Marginal ranking for gender (main identity ranking question).
  }
  \item{race}{
    Marginal ranking for race (main identity ranking question).
  }
  \item{anc_identity}{Full ranking profile for the anchor ranking question.}
  \item{household}{Marginal ranking for household (anchor question).}
  \item{neighborhood}{
    Marginal ranking for neighborhood (anchor question).
  }
  \item{city}{Marginal ranking for city (anchor question).}
  \item{state}{Marginal ranking for state (anchor question).}
  \item{anc_correct_identity}{
    Whether the respondent answered the anchor questions correctly.
    This is a binary variable that 1 if the respondent correctly answers
    the anchor ranking question and 0 if otherwise.
  }
  \item{app_identity_recorded}{
    Recorded responses for the main identity ranking question.
  }
  \item{anc_identity_recorded}{
    Recorded responses for the anchor ranking question.
  }
  \item{app_identity_row_rnd}{
    The order in which the items were randomly presented for the
    respondent in the main ranking question.
  }
  \item{anc_identity_row_rnd}{
    The order in which the items were randomly presented for the
    respondent in the anchor ranking question.
  }
  \item{ranking}{
    Pasted full ranking profile reconstructed from the marginal ranking
    columns. In this dataset, it matches \code{app_identity}.
  }
}
}
\source{
<https://github.com/sysilviakim/ranking_error>
}
\usage{
identity_w
}
\description{
This data is the `results` element returned by applying
`imprr_weights()` to `identity`. It adds two columns to the original
`identity` data: `weights`, the estimated inverse probability weights, and
`ranking`, the pasted full ranking profile.
}
\keyword{datasets}
