% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_to_rank.R
\name{item_to_rank}
\alias{item_to_rank}
\title{Return Rankings with Items as Columns}
\usage{
item_to_rank(
  item_rank,
  format_input = "ordering",
  reference = NULL,
  long = FALSE
)
}
\arguments{
\item{item_rank}{A data frame with rankings as columns,
with items being ranked as cell values.}

\item{format_input}{Character string indicating the format of the data input,
namely "ordering" or "ranking".
Used for \code{PLMIX::rank_ord_switch}.}

\item{reference}{A character vector of item names to be used for renaming
the columns. If not specified, will use the first 26 letters of the alphabet.
Default is `NULL`.}

\item{long}{Whether to return the output in a long data format.
Default is `FALSE`.}
}
\value{
A data frame with items that are being ranked as columns,
with rankings in cell values.
}
\description{
This function takes a ranking dataset with rankings as columns and
returns a dataset with items as columns and rankings as cell values.
This function is useful for converting rankings to a format that allows for
average ranking calculations.
}
\examples{
true_pref <- data.frame(
  first = c("b", "a", "c"),
  second = c("c", "b", "b"),
  third = c("a", "c", "a")
)
item_to_rank(true_pref)
item_to_rank(true_pref, long = TRUE)

}
