% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ordinal_seq.R
\name{ordinal_seq}
\alias{ordinal_seq}
\title{Generate an Ordinal Sequence from a Number}
\usage{
ordinal_seq(length)
}
\arguments{
\item{length}{The length of the ordinal sequence to generate.
It should be a numeric value of length 1.}
}
\value{
A vector of ordinal strings.
}
\description{
This function generates an ordinal sequence of an arbitrary length. For
example, if the length is 3, the function will return the vector
\code{c("1st", "2nd", "3rd")}.
}
\examples{
ordinal_seq(11)

}
