% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permn_augment.R
\name{permn_augment}
\alias{permn_augment}
\title{Augmenting Permutation Patterns}
\usage{
permn_augment(tab, J = 4)
}
\arguments{
\item{tab}{A table of observed permutation patterns.}

\item{J}{The length of the reference choice set.}
}
\value{
A table of observed permutation patterns augmented with all possible
permutation patterns.
}
\description{
In some distribution of ranking data, not all possible permutation patterns
may be realized due to the sample size or skewed distribution of preferences.
}
\details{
This function augments the given table with all possible observed
permutation patterns with a frequency of zero for unrealized patterns.
}
\examples{
tab <- table(c(rep("123", 100), rep("321", 50)))
permn_augment(tab, J = 3)

}
