% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_avg_ranking.R
\name{plot_average_rank}
\alias{plot_average_rank}
\title{Plot Average Rank Results}
\usage{
plot_average_rank(data, qoi_filter = "average rank", xlab = NULL, ylab = "")
}
\arguments{
\item{data}{The results data from the `imprr_direct` function.
If an external data frame, make sure that the column names
are the same as the output from the `imprr_direct` function.}

\item{qoi_filter}{The quantity of interest (QOI) to filter for.
Defaults to "average rank".}

\item{xlab}{The x-axis label. Defaults to NULL.
If you'd like it to be empty, specify an empty string.}

\item{ylab}{The y-axis label. Defaults to an empty string.}
}
\value{
A ggplot object.
}
\description{
This function takes the output from the `imprr_direct` function
and plots the average rank results with confidence intervals.
As long as the mean and the confidence intervals are provided,
this function will plot other quantities of interest such as
marginal, pairwise, top-k rankings.
}
