% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_ranking.R
\name{plot_dist_ranking}
\alias{plot_dist_ranking}
\title{Plot the Distribution of Rankings Over the Permutation Space}
\usage{
plot_dist_ranking(
  tab,
  x = "ranking",
  y = "prop",
  ylim = 0.315,
  fill = "firebrick4",
  xlab = "Recorded Responses",
  family = NULL,
  vjust = -0.5,
  size = 3,
  linetype = "dashed",
  h_color = "black",
  h_alpha = 0.5
)
}
\arguments{
\item{tab}{A table in which the frequencies of ranking patterns are recorded.}

\item{x}{Name of the column that contains permutation patterns.}

\item{y}{Name of the column that contains the proportion of ranking
permutations.}

\item{ylim}{The upper limit of the y-axis.}

\item{fill}{The color of the bars.}

\item{xlab}{The label of the x-axis.
Defaults to "Recorded Responses".}

\item{family}{The font family of the text.}

\item{vjust}{The vertical justification of the text.}

\item{size}{The size of the text in `geom_text`.}

\item{linetype}{The linetype in `geom_hline`.}

\item{h_color}{The color in `geom_hline`.}

\item{h_alpha}{The transparency in `geom_hline`.}
}
\value{
A ggplot2 object.
}
\description{
This function takes a table in which the frequencies of ranking patterns
are recorded and plots it over the permutation space of rankings, using
the \code{ggplot2} package.
}
\examples{
tab <- lapply(combinat::permn(seq(3)), paste0, collapse = "") |>
  sample(30, replace = TRUE) |>
  unlist() |>
  table() |>
  table_to_tibble()
plot_dist_ranking(tab, ylim = 0.5)

}
