% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_longer.R
\name{rank_longer}
\alias{rank_longer}
\title{Turn Column(s) Recording Rankings in Wide Format into a Long Data Format}
\usage{
rank_longer(x, cols = NULL, id = NULL, reference = NULL)
}
\arguments{
\item{x}{A data frame in wide format with columns recording rankings.}

\item{cols}{A character vector of column names that record rankings.
If there are multiple columns, the order of the columns should be the same
as the order of the reference choice set. If there is a single column, the
order in which the numbers appear in respondent-level character response
should be the same as the order of the reference set.}

\item{id}{The column that uniquely identify the respondent.}

\item{reference}{If you wish to specify the reference choice set,
you can provide a character vector.}
}
\value{
A data frame in long format with columns recording rankings.
The first column is the id variable that has been pre-specified.
The second and third columns record what item is being ranked.
The final column records the ranking of the item.
}
\description{
This function takes a data frame in wide format with columns recording
rankings into a long data format.
}
\details{
If the data frame has more than one columns specified in the `cols` argument,
they will be translated as the first, second, third, etc. items in the
reference choice set. For example, if the first column records 2, the second
column records 1, and the third column records 3, then the function will
interpret that this respondent prefers the second item the most,
then the first item, then the third item.

If the data frame has only one column specified in the `cols` argument, it
will be parsed by character length. For example, if the column records "213",
then the function will interpret that this respondent prefers the second item
the most, then the first, and then the third item.

Currently, this function depends on `tidyverse` functions.
Eventually, a `data.table` option will be added for large datasets.
}
\examples{
x <- data.frame(
  apple = c(2, 1, 3),
  orange = c(1, 3, 2),
  banana = c(3, 2, 1)
)

rank_longer(x)

y <- data.frame(
  id = c("Bernie", "Yuki", "Silvia"),
  rank = c("123", "321", "213")
)
rank_longer(y, cols = "rank", id = "id")
rank_longer(
  y, cols = "rank", id = "id",
  reference = c("Baseball", "Basketball", "Football")
)

}
