% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_to_tibble.R
\name{table_to_tibble}
\alias{table_to_tibble}
\title{Turn the Frequency Table into a Tibble or Data Frame}
\usage{
table_to_tibble(tab, tibble = TRUE)
}
\arguments{
\item{tab}{A frequency table.}

\item{tibble}{A logical value indicating whether the output should be
a tibble or data frame. Defaults to TRUE.}
}
\value{
A tibble or data frame, depending on the `tibble` argument.
}
\description{
This function converts a frequency table to a tibble or data frame.
It also creates a proportion variable as well as the frequency variable.
}
