% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unbiased_correct_prop.R
\name{unbiased_correct_prop}
\alias{unbiased_correct_prop}
\title{Unbiased Estimator of the Proportion of Random and Non-random Responses}
\usage{
unbiased_correct_prop(mean_c, J)
}
\arguments{
\item{mean_c}{This is the raw proportion of the correct answers.}

\item{J}{The number of items to rank order.}
}
\value{
A number between 0-1.
}
\description{
This function computes the unbiased proportion of *correct* answers after
adjusting for the possibility that the respondent may have randomly guessed
the correct answer. The function is based on the formula provided by
Proposition 1 of Atsusaka and Kim (2024).
}
\examples{
unbiased_correct_prop(0.7, 3)

}
