% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprr_weights.R
\name{wtd.table}
\alias{wtd.table}
\title{Copied from questionr::wtd.table
Not exported}
\usage{
wtd.table(
  x,
  y = NULL,
  weights = NULL,
  digits = 3,
  normwt = FALSE,
  useNA = c("no", "ifany", "always"),
  na.rm = TRUE,
  na.show = FALSE,
  exclude = NULL
)
}
\description{
Copied from questionr::wtd.table
Not exported
}
