% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_rank.R
\name{avg_rank}
\alias{avg_rank}
\title{Compute the Average Rank of All Items}
\usage{
avg_rank(x, rankings, items = NULL, type = "long")
}
\arguments{
\item{x}{A data frame that contains rankings of items.}

\item{rankings}{The name of the column that contains the rankings.}

\item{items}{The name of the column that contains the items' names, or,
in case of a wide file, the item names in the reference choice set.}

\item{type}{The type of the data frame.
It defaults to \code{"long"}, which means that the data frame is in
the long format, presumably generated by \code{rank_longer()}.
If the data frame is in the wide format, it should be set to \code{"wide"}.
This only accepts a data frame with a single ranking variable that
contains rankings, and the rankings should be in the form of a string
such as \code{"123"}.}
}
\value{
A data frame with the average rank of each item in the
reference choice set.
}
\description{
This function calculates the average rank for the data frame that contains
ranking data. It can be used for both long- and wide-type data frames.
}
\examples{
x <- data.frame(
  id = c("Bernie", "Yuki", "Silvia"),
  rank = c("123", "321", "213")
)
avg_rank(x, "rank", type = "wide")
avg_rank(x, "rank", type = "wide", items = c("Money", "Power", "Respect"))

y <- data.frame(rank = c("123", "321", "213"))
avg_rank(y, "rank", type = "wide")
z <- rank_longer(
  y,
  cols = "rank", id = "id",
  reference = c("Money", "Power", "Respect")
)
avg_rank(z, "ranking", items = "item_name")

}
