% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_rank.R
\name{avg_rank}
\alias{avg_rank}
\title{Compute the Average Rank of All Items}
\usage{
avg_rank(
  x,
  rankings = NULL,
  items = NULL,
  long = FALSE,
  raw = TRUE,
  weight = NULL,
  round = NULL
)
}
\arguments{
\item{x}{A data frame that contains rankings of items.}

\item{rankings}{The name of the column that contains the rankings.
Defaults to NULL, which means that the function will look for a data frame
with two columns, "item" and "rank".}

\item{items}{The name of the column that contains the items' names, or,
in case of a wide file, the item names in the reference choice set.
Defaults to NULL.}

\item{long}{The type of the data frame. Defaults to `FALSE`.
It `TRUE`, which means that the data frame is in
the long format, it is presumed to be generated by \code{rank_longer()}.
If the data frame is in the wide format, it should be set to `FALSE`.
This only accepts a data frame with a single ranking variable that
contains rankings, and the rankings should be in the form of a string
such as \code{"123"}.}

\item{raw}{If \code{TRUE}, the function will return the raw average rank.
If \code{FALSE}, the function will return the average rank after correcting
based on the IPW estimator. Defaults to \code{TRUE}.}

\item{weight}{The name of the column that contains the weights for the IPW
estimator. Defaults to \code{NULL}.}

\item{round}{The number of decimal places to round the output to.
Defaults to \code{NULL}.}
}
\value{
A data frame with the average rank of each item in the
reference choice set.
}
\description{
This function calculates the average rank for the data frame that contains
ranking data. It can be used for both long- and wide-type data frames.
}
\examples{
x <- data.frame(
  id = c("Bernie", "Yuki", "Silvia"),
  rank = c("123", "321", "213")
)
avg_rank(x, "rank")
avg_rank(x, "rank", items = c("Money", "Power", "Respect"))

y <- data.frame(rank = c("123", "321", "213"))
avg_rank(y, "rank")
z <- rank_longer(
  y,
  cols = "rank", id = "id",
  reference = c("Money", "Power", "Respect")
)
avg_rank(z, "ranking", items = "item_name", long = TRUE)

## Example output from item_to_rank
x <- data.frame(
  item = c("a", "b", "c", "a", "b", "c", "a", "b", "c"),
  rank = c(3L, 1L, 2L, 1L, 2L, 3L, 3L, 2L, 1L)
)
avg_rank(x, long = TRUE)

}
