% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{identity_w}
\alias{identity_w}
\title{Identity-ranking data with estimated weights based on inverse probability weighting}
\format{
## `identity_w`
A data frame with 1,082 rows and 12 columns:
\describe{
  \item{w}{Estimated weight based on inverse probability weighting.}
  \item{ranking}{Ranking pattern that the respondent provided.}
  \item{app_party}{Marginal ranking for party (main identity ranking question).}
  \item{app_religion}{Marginal ranking for religion (main identity ranking question).}
  \item{app_gender}{Marginal ranking for gender (main identity ranking question).}
  \item{app_race}{Marginal ranking for race (main identity ranking question).}
  \item{anc_house}{Marginal ranking for household (anchor question).}
  \item{anc_neighborhood}{Marginal ranking for neighborhood (anchor question).}
  \item{anc_city}{Marginal ranking for city (anchor question).}
  \item{anc_state}{Marginal ranking for state (anchor question).}
  \item{anc_correct_identity}{Whether the respondent answered the anchor questions correctly. This is a binary variable that 1 if the respondent correctly answers the anchor ranking question and 0 if otherwise.}
  \item{s_weight}{Survey weight.}
}
}
\source{
<https://github.com/sysilviakim/ranking_error>
}
\usage{
identity_w
}
\description{
This data has two extra columns from the original `identity` data:
`w` and `ranking`, which are respectively the estimated weights based on
inverse probability weighting and the ranking pattern that the respondent
provided, united into a single column.
}
\keyword{datasets}
