% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprr_direct.R
\name{imprr_direct}
\alias{imprr_direct}
\title{Bias-correct the Distribution of Ranking Permutations
Using a Paired Anchor Question and Produce De-biased Average Rankings}
\usage{
imprr_direct(
  data,
  J = NULL,
  main_q,
  anchor_q,
  anc_correct,
  anc_correct_pattern = NULL,
  n_bootstrap = 200,
  seed = 123456
)
}
\arguments{
\item{data}{The input dataset with ranking data.}

\item{J}{The number of items in the ranking question. Defaults to NULL,
in which case it will be inferred from the data.}

\item{main_q}{Column name for the main ranking question to be analyzed.}

\item{anchor_q}{Column name for the paired anchor question.}

\item{anc_correct}{Indicator for passing the anchor question.}

\item{anc_correct_pattern}{The correct pattern to pass the anchor question
filter, given the reference set. Defaults to NULL.
If NULL, it will taken on a J-length string with a natural progression of
numbers, such as "123", "1234", "1234567", and so on.}

\item{n_bootstrap}{Number of bootstraps. Defaults to 200.}

\item{seed}{Seed for \code{set.seed} for reproducibility.}
}
\value{
A list.
}
\description{
This function implements the bias correction of the ranking
distribution using a paired anchor question.
}
