% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprr_direct.R
\name{imprr_direct}
\alias{imprr_direct}
\title{Implements Plug-in Bias-Corrected Estimators for Ranking Data}
\usage{
imprr_direct(
  data,
  J = NULL,
  main_q,
  anc_correct,
  population = "non-random",
  assumption = "contaminated",
  n_bootstrap = 200,
  seed = 123456,
  weight = NULL,
  verbose = FALSE
)
}
\arguments{
\item{data}{The input dataset with ranking data.}

\item{J}{The number of items in the ranking question. Defaults to NULL,
in which case it will be inferred from the data, only if the column for
`main_q` exists in the data.}

\item{main_q}{Column name for the main ranking question to be analyzed.
Using this argument, the function automatically looks for columns with
marginal rankings. For example, if `main_q` is `app_identity`, the function
looks for `app_identity_1`, `app_identity_2`, `app_identity_3`, and so on,
with an underbar separator followed by numbers.}

\item{anc_correct}{Indicator for passing the anchor question.}

\item{population}{Choice of the target population out of
non-random respondents (default) or all respondents.}

\item{assumption}{Choice of the identifying assumption if `population` is set to all}

\item{n_bootstrap}{Number of bootstraps. Defaults to 200.}

\item{seed}{Seed for \code{set.seed} for reproducibility.}

\item{weight}{A vector of weights. Defaults to NULL.}

\item{verbose}{Indicator for verbose output. Defaults to FALSE.}
}
\value{
A list.
}
\description{
This function implements the bias correction of the ranking
distribution using a paired anchor question.
}
