% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dist_ranking.R
\name{plot_dist_ranking}
\alias{plot_dist_ranking}
\title{Plot the Distribution of Rankings Over the Permutation Space}
\usage{
plot_dist_ranking(
  tab,
  x = "ranking",
  y = "prop",
  ylim = 0.315,
  fill = "firebrick4",
  xlab = "Recorded Rankings",
  family = "Sans",
  vjust = -0.5,
  size = 3
)
}
\arguments{
\item{tab}{A table in which the frequencies of ranking patterns are recorded.}

\item{x}{Name of the column that contains permutation patterns.}

\item{y}{Name of the column that contains the proportion of ranking
permutations.}

\item{ylim}{The upper limit of the y-axis.}

\item{fill}{The color of the bars.}

\item{xlab}{The label of the x-axis.}

\item{family}{The font family of the text.}

\item{vjust}{The vertical justification of the text.}

\item{size}{The size of the text in `geom_text`.}
}
\value{
A ggplot2 object.
}
\description{
This function takes a table in which the frequencies of ranking patterns
are recorded and plots it over the permutation space of rankings, using
the `ggplot2` package.
}
