% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankingQ-package.R
\docType{package}
\name{rankingQ-package}
\alias{rankingQ}
\alias{rankingQ-package}
\title{rankingQ: Design-Based Methods for Improving Ranking Questions}
\description{
\if{html}{\figure{logo.png}{
options: style='float: right' alt='logo' width='120'}}

Implements design-based bias correction methods that address
measurement errors due to random responses in ranking survey questions.
Based on Atsusaka and Kim (2025) \doi{10.1017/pan.2024.33}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/sysilviakim/rankingQ}
  \item \url{http://sysilviakim.com/rankingQ/}
  \item Report bugs at \url{https://github.com/sysilviakim/rankingQ/issues}
}

}
\author{
\strong{Maintainer}: Seo-young Silvia Kim
\email{sskim.research@gmail.com}
(\href{https://orcid.org/0000-0002-8801-9210}{ORCID})

Authors:
\itemize{
  \item Yuki Atsusaka
  \email{yuki.atsusaka@gmail.com}
  (\href{https://orcid.org/0000-0001-5365-1876}{ORCID})
}

}
\keyword{internal}
