% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_avg.R
\name{stratified_avg}
\alias{stratified_avg}
\title{Stratified Estimate of Average Ranks}
\usage{
stratified_avg(
  data,
  var_stratum,
  J = NULL,
  main_q,
  anc_correct = NULL,
  labels = NULL,
  seed = 1234,
  weight = NULL,
  n_bootstrap = 200,
  ipw = FALSE,
  verbose = FALSE,
  p_random = NULL
)
}
\arguments{
\item{data}{A data frame containing the ranking data as well as the
stratifying variable.}

\item{var_stratum}{The name of the stratifying variable.}

\item{J}{The number of items in the ranking question. Defaults to NULL,
in which case it will be inferred from the data.}

\item{main_q}{Column name for the main ranking question to be analyzed.}

\item{anc_correct}{Optional indicator for passing the anchor question.
If `NULL`, `p_random` is used when supplied; otherwise the function
defaults to `p_random = 0` and applies no correction.}

\item{labels}{A vector of labels for the items being ranked.
Defaults to NULL.}

\item{seed}{Seed for \code{set.seed} for reproducibility.}

\item{weight}{Either a numeric vector of weights with length `nrow(data)`
or the name of a weight column in `data`. Defaults to `NULL`.}

\item{n_bootstrap}{Number of bootstraps. Defaults to 200.}

\item{ipw}{Indicator for using inverse probability weighting. Defaults to
FALSE, in which case direct bias estimation will be employed.}

\item{verbose}{Indicator for verbose output. Defaults to FALSE.}

\item{p_random}{Optional fixed proportion of random/inattentive respondents.
When supplied, this overrides `anc_correct`.}
}
\value{
A data frame with the bootstrap-estimated average ranks.
}
\description{
This function estimates the average ranks based on
stratification.
}
