% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_avg.R
\name{stratified_avg}
\alias{stratified_avg}
\title{Stratified Estimate of Average Ranks}
\usage{
stratified_avg(
  data,
  var_stratum,
  J = NULL,
  main_q,
  anc_correct,
  labels = NULL,
  seed = 1234,
  weight = NULL,
  n_bootstrap = 200,
  ipw = FALSE
)
}
\arguments{
\item{data}{A data frame containing the ranking data as well as the
stratifying variable.}

\item{var_stratum}{The name of the stratifying variable.}

\item{J}{The number of items in the ranking question. Defaults to NULL,
in which case it will be inferred from the data.}

\item{main_q}{Column name for the main ranking question to be analyzed.}

\item{anc_correct}{Indicator for passing the anchor question.}

\item{labels}{A vector of labels for the items being ranked.
Defaults to NULL.}

\item{seed}{Seed for \code{set.seed} for reproducibility.}

\item{weight}{A vector of weights. Defaults to NULL.}

\item{n_bootstrap}{Number of bootstraps. Defaults to 200.}

\item{ipw}{Indicator for using inverse probability weighting. Defaults to
FALSE, in which case direct bias estimation will be employed.}
}
\value{
A data frame with the bootstrap-estimated average ranks.
}
\description{
This function estimates the average ranks based on
stratification.
}
