% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniformity_test.R
\name{uniformity_test}
\alias{uniformity_test}
\title{Uniformity Test for Ranking Patterns}
\usage{
uniformity_test(data, var = NULL)
}
\arguments{
\item{data}{The input dataset with ranking data.}

\item{var}{The variable within \code{data} to be used in the test.
Defaults to NULL.}
}
\value{
A chi-square test result.
}
\description{
This function implements the uniformity test for
ranking permutation patterns documented in Atsusaka and Kim (2024).
}
