% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/imprr_weights.R
\name{wtd.table}
\alias{wtd.table}
\title{Weighted one-way and two-way frequency tables.}
\usage{
wtd.table(
  x,
  y = NULL,
  weights = NULL,
  digits = 3,
  normwt = FALSE,
  useNA = c("no", "ifany", "always"),
  na.rm = TRUE,
  na.show = FALSE,
  exclude = NULL
)
}
\arguments{
\item{x}{a vector}

\item{y}{another optional vector for a two-way frequency table. Must be the same length as x}

\item{weights}{vector of weights, must be the same length as x}

\item{digits}{Number of significant digits.}

\item{normwt}{if TRUE, normalize weights so that the total weighted count is the same as the unweighted one}

\item{useNA}{whether to include NA values in the table}

\item{na.rm}{(deprecated) if TRUE, remove NA values before computation}

\item{na.show}{(deprecated) if TRUE, show NA count in table output}

\item{exclude}{values to remove from x and y. To exclude NA, use na.rm argument.}
}
\description{
Copied from questionr::wtd.table. Not exported
}
